DPTSW-19618: [linlon-dp] Set AFBC 32x8 to the highest priority in the supported modifier lists
For AFBC buffers with width exceeding 2048, only the 32x8 block size is supported.

Because some display services select the first modifier supported by the plane,
the AFBC_32x8 (_TILED | _SC | _SPLIT | _SPARSE | _YTR) modifier is moved to the
top of the supported modifiers list.

This gives it the highest priority and ensures proper handling of wide AFBC buffers.

Signed-off-by: cunyuan.liu <cunyuan.liu@cixtech.com>
Change-Id: I2c6bc8b62247ab84b01dea89ac8b6f8ae039bb90